refactor: tidy up pages component, use ISR + fallback for homepage
