+++ src/components/pages/home/types.ts
@@ -1,0 +1,8 @@
+import { MovieListResponse } from "services/tmdb/movie/list/types";
+
+export type HomePageProps = {
+  popularFallbackData: MovieListResponse;
+  nowPlayingFallbackData: MovieListResponse;
+  topRatedFallbackData: MovieListResponse;
+  upcomingFallbackData: MovieListResponse;
+};
